added solutions for tasks 4 and 5 added input validation checks fixed shortcomings in code design

+++ task01javabasic/src/main/java/by/zakharanka/task01javabasic/service/ExpressionService.java
@@ -11,7 +11,7 @@ public class ExpressionService {
         return ((b + Math.sqrt((b * b) + 4 * a * c)) / (2 * a)) - (a * a * a * c) + (1 / b * b);
     }
 
-    public double milkTask(Data<Double> data){
+    public double milkTask(Data<Integer> data){
         double n = data.getData(0);
         double m = data.getData(1);
         return ((80 / n) + 12) * m;
